Re #27664 Renamed StateRunData and fold in extra branch
Renames RunDataStateBuilder to StateRunDataBuilder to match the naming
convention of all other state objects. Also removes an extra branch
within StateRunData as pointed out in a PR review.
