+++ scripts/test/SANS/state/CMakeLists.txt
@@ -15,6 +15,7 @@ set(TEST_PY_FILES
     slice_event_test.py
     save_test.py
     state_test.py
+    StateRunDataBuilderTest.py
     wavelength_test.py
     wavelength_and_pixel_adjustment_test.py)
